backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
python-jose[cryptography]==3.3.0
requests==2.32.3
